more map stuff
