+++ src/interface/src/app/map/map.layers.ts
@@ -1,0 +1,77 @@
+import { Map } from '../types';
+import * as L from 'leaflet';
+
+export function addClonedLayerToMap(map: Map, layer: L.Layer) {
+  const originalId = L.Util.stamp(layer);
+
+  // Hacky way to clone, but it removes the reference to the origin layer
+  const clonedLayer = L.geoJson((layer as L.Polygon).toGeoJSON()).setStyle({
+    color: '#ffde9e',
+    fillColor: '#ffde9e',
+    weight: 5,
+  });
+  map.clonedDrawingRef?.addLayer(clonedLayer);
+  map.drawnPolygonLookup![originalId] = clonedLayer;
+}
+
+export function removeClonedLayer(
+  map: Map,
+  layer: L.Layer,
+  deleteOriginal: boolean
+) {
+  const originalPolygonKey = L.Util.stamp(layer);
+  const clonedPolygon = map.drawnPolygonLookup![originalPolygonKey];
+  map.clonedDrawingRef!.removeLayer(clonedPolygon);
+  if (deleteOriginal) {
+    delete map.drawnPolygonLookup![originalPolygonKey];
+  }
+}
+
+export function addRegionLayer(map: Map, boundary: any) {
+  // Add corners of the map to invert the polygon
+  if (map.regionLayerRef) {
+    map.regionLayerRef?.remove();
+  }
+  map.regionLayerRef = createRegionLayer(boundary);
+  map.regionLayerRef.addTo(map.instance!);
+}
+
+export function showRegionLayer(map: Map) {
+  if (map.regionLayerRef) {
+    map.regionLayerRef.setStyle({ opacity: 1 });
+  }
+}
+
+export function hideRegionLayer(map: Map) {
+  if (map.regionLayerRef) {
+    map.regionLayerRef.setStyle({ opacity: 0 });
+  }
+}
+
+function createRegionLayer(boundaries: GeoJSON.GeoJSON) {
+  return L.geoJSON(boundaries, {
+    style: (_) => ({
+      color: '#93b3ff',
+      weight: 4,
+      opacity: 0,
+      fillColor: '#000000',
+      fillOpacity: 0,
+    }),
+  });
+}
+
+export function createDrawingLayer(
+  planningAreaData: GeoJSON.GeoJSON,
+  color?: string,
+  opacity?: number
+) {
+  return L.geoJSON(planningAreaData, {
+    pane: 'overlayPane',
+    style: {
+      color: color ?? '#3367D6',
+      fillColor: color ?? '#3367D6',
+      fillOpacity: opacity ?? 0.1,
+      weight: 7,
+    },
+  });
+}
